Update to geos 3.8.1
Fixes #10

+++ Sources/geos/include/geos/noding/snapround/MCIndexPointSnapper.h
@@ -19,21 +19,22 @@
 #ifndef GEOS_NODING_SNAPROUND_MCINDEXPOINTSNAPPER_H
 #define GEOS_NODING_SNAPROUND_MCINDEXPOINTSNAPPER_H
 
+#include <cstddef>
 #include <geos/export.h>
 
 #include <geos/inline.h>
 
 // Forward declarations
 namespace geos {
-	namespace index {
-		class SpatialIndex;
-	}
-	namespace noding {
-		class SegmentString;
-		namespace snapround {
-			class HotPixel;
-		}
-	}
+namespace index {
+class SpatialIndex;
+}
+namespace noding {
+class SegmentString;
+namespace snapround {
+class HotPixel;
+}
+}
 }
 
 namespace geos {
@@ -41,8 +42,8 @@ namespace noding { // geos::noding
 namespace snapround { // geos::noding::snapround
 
 /** \brief
- * "Snaps" all {@link SegmentString}s in a {@link SpatialIndex} containing
- * {@link MonotoneChain}s to a given {@link HotPixel}.
+ * "Snaps" all [SegmentStrings](@ref SegmentString) in a [SpatialIndex](@ref index::SpatialIndex) containing
+ * [MonotoneChains](@ref index::chain::MonotoneChain) to a given HotPixel.
  *
  */
 class GEOS_DLL MCIndexPointSnapper {
@@ -50,34 +51,36 @@ class GEOS_DLL MCIndexPointSnapper {
 public:
 
 
-	MCIndexPointSnapper(index::SpatialIndex& nIndex)
-		:
-		index(nIndex)
-	{}
+    MCIndexPointSnapper(index::SpatialIndex& nIndex)
+        :
+        index(nIndex)
+    {}
 
-	/**
-	 * Snaps (nodes) all interacting segments to this hot pixel.
-	 * The hot pixel may represent a vertex of an edge,
-	 * in which case this routine uses the optimization
-	 * of not noding the vertex itself
-	 *
-	 * @param hotPixel the hot pixel to snap to
-	 * @param parentEdge the edge containing the vertex,
-	 *        if applicable, or <code>null</code>
-	 * @param vertexIndex the index of the vertex, if applicable, or -1
-	 * @return <code>true</code> if a node was added for this pixel
-	 */
-	bool snap(HotPixel& hotPixel, SegmentString* parentEdge,
-			unsigned int vertexIndex);
+    /** \brief
+     * Snaps (nodes) all interacting segments to this hot pixel.
+     *
+     * The hot pixel may represent a vertex of an edge, in which case this
+     * routine uses the optimization of not noding the vertex itself.
+     *
+     * @param hotPixel the hot pixel to snap to
+     * @param parentEdge the edge containing the vertex,
+     *                   if applicable, or `null`
+     * @param vertexIndex the index of the vertex, if applicable, or -1
+     * @return `true` if a node was added for this pixel
+     */
+    bool snap(HotPixel& hotPixel, SegmentString* parentEdge,
+              size_t vertexIndex);
 
-	bool snap(HotPixel& hotPixel) {
-		return snap(hotPixel, nullptr, 0);
-	}
+    bool
+    snap(HotPixel& hotPixel)
+    {
+        return snap(hotPixel, nullptr, 0);
+    }
 
 
 private:
 
-	index::SpatialIndex& index;
+    index::SpatialIndex& index;
 
     // Declare type as noncopyable
     MCIndexPointSnapper(const MCIndexPointSnapper& other) = delete;
